requirements.txt
asgiref==3.6.0
Django==4.2.1
djangorestframework==3.14.0
isort==5.12.0
pytz==2023.3
sqlparse==0.4.4
python-dotenv
drf-spectacular
requests==2.26.0
redis
celery
django-celery-beat
flower
